clean-up stream response constructors
